infinitylists: A Shiny-Based Application for Generating Place-Based Species Lists

This shiny-based application allows users to extract species occurrence records from the Atlas of Living Australia or GBIF and generate a species list for a defined area. All records associated with either a physical voucher (stored in herbaria or museums) or a photographic or audio voucher (stored in iNaturalist) are extracted. For each species within the defined area, the application will return voucher type, number of vouchers, date of the most recent voucher, spatial coordinates, voucher location, and the voucher collector. Records are displayed both in a table and on a map, and are downloadable as a CSV.

See also

Author

Maintainer: Thomas Mesaglio thomasmesaglio@gmail.com (ORCID) [copyright holder]

Authors:

Will Cornwell w.cornwell@unsw.edu.au (ORCID)

Fonti Kar f.kar@unsw.edu.au (ORCID)